requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
python-dotenv==1.0.0
supabase==1.0.4
gotrue==1.0.4
gunicorn==21.2.0
bcrypt==4.1.2